net: Moved UDP files to own directory
