+++ api/net/inet.hpp
@@ -29,13 +29,13 @@
 #include "conntrack.hpp"
 
 #include "ip4/ip4.hpp"
-#include "ip4/udp.hpp"
 #include "ip4/icmp4.hpp"
 #include "ip4/arp.hpp"
 #include "ip6/ip6.hpp"
 #include "ip6/icmp6.hpp"
 #include "dns/client.hpp"
 #include "tcp/tcp.hpp"
+#include "udp/udp.hpp"
 #include "super_stack.hpp"
 
 namespace net {
